test: separate out faucet.test.ts
